Multiple Sets of hint fades, incorrect hint reporting.
tracking if a row is correct and if it is completely filled in separately. Now we fade the hints differently for 4 possible states:
correct and filled in (fade the most)
correct but not marked totally empty (faded less)
incorrect but still empty spots (still in progress)
incorrect and completely filled with marks (incorrect).

maybe want to put some of this behaviour behind an option?