Fix picross crashing on victory
previous memory fixes were freeing up the memory of level select, but that happens after the picross level is initialized, so it would get initialized properly, then free the victory wsg, so it couldn't be displayed when you won. I now copy the completed puzzle instead of referencing it by pointer.